--- patches/expo-modules-core+3.0.1.patch
@@ -1,71 +1,0 @@
-diff --git a/node_modules/expo-modules-core/ios/Core/Views/SwiftUI/Convertibles+SwiftUI.swift b/node_modules/expo-modules-core/ios/Core/Views/SwiftUI/Convertibles+SwiftUI.swift
-index 71ef927..16fc90c 100644
---- a/node_modules/expo-modules-core/ios/Core/Views/SwiftUI/Convertibles+SwiftUI.swift
-+++ b/node_modules/expo-modules-core/ios/Core/Views/SwiftUI/Convertibles+SwiftUI.swift
-@@ -1,6 +1,7 @@
- // Copyright 2024-present 650 Industries. All rights reserved.
- 
- import SwiftUI
-+import OSLog
- 
- extension Color: Convertible {
-   public static func convert(from value: Any?, appContext: AppContext) throws -> Color {
-@@ -72,6 +73,58 @@ extension Color: Convertible {
-       return nil
-     }
-   }
-+
-+  #if os(tvOS)
-+  private static let _systemGray6: Color = Color(.systemGray.withAlphaComponent(0.5))
-+  private static let _systemGray5: Color = Color(.systemGray.withAlphaComponent(0.6))
-+  private static let _systemGray4: Color = Color(.systemGray.withAlphaComponent(0.7))
-+  private static let _systemBackground: Color = .clear
-+  private static let _systemGroupedBackground: Color = .white
-+  private static let _secondaryLabel: Color = .secondary
-+  private static let _secondarySystemBackground: Color = Color(.systemGray.withAlphaComponent(0.1))
-+  private static let _secondarySystemGroupedBackground: Color = Color(.systemGray.withAlphaComponent(0.2))
-+  #else
-+  private static let _systemGray6: Color = Color(.systemGray6)
-+  private static let _systemGray5: Color = Color(.systemGray5)
-+  private static let _systemGray4 = Color(.systemGray4)
-+  private static let _systemBackground: Color = .systemBackground
-+  private static let _systemGroupedBackground: Color = .systemGroupedBackground
-+  private static let _secondaryLabel: Color = .secondaryLabel
-+  private static let _secondarySystemBackground: Color = .secondarySystemBackground
-+  private static let _secondarySystemGroupedBackground: Color = ._secondarySystemGroupedBackground
-+  #endif
-+
-+  public enum expoSystemColors {
-+    case systemBackground
-+    case systemGroupedBackground
-+    case secondaryLabel
-+    case secondarySystemBackground
-+    case secondarySystemGroupedBackground
-+    case systemGray4
-+    case systemGray5
-+    case systemGray6
-+
-+    public var color: Color {
-+      switch(self) {
-+      case .secondaryLabel:
-+        return _secondaryLabel
-+      case .secondarySystemBackground:
-+        return _secondarySystemBackground
-+      case .secondarySystemGroupedBackground:
-+        return _secondarySystemGroupedBackground
-+      case .systemBackground:
-+        return _systemBackground
-+      case .systemGroupedBackground:
-+        return _systemGroupedBackground
-+      case .systemGray4:
-+        return _systemGray4
-+      case .systemGray5:
-+        return _systemGray5
-+      case .systemGray6:
-+        return _systemGray6
-+      }
-+    }
-+  }
- }
- 
- extension UnitPoint: Convertible {
